Refactor customer credit facilities entrypoint
